altered product_quantity datatype

+++ backend/database/schema.sql
@@ -20,7 +20,7 @@ CREATE TABLE todo_item (
     item_id INT PRIMARY KEY AUTO_INCREMENT,
     list_id INT NOT NULL,
     product_name VARCHAR(100) NOT NULL,
-    product_quantity DECIMAL(10,2) NOT NULL,
+    product_quantity INT NOT NULL,
     is_completed BOOLEAN DEFAULT FALSE,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     FOREIGN KEY (list_id) REFERENCES grocery_list(list_id) ON DELETE CASCADE
